+++ packages/terra-worklist-data-grid/src/WorklistDataGrid.jsx
@@ -1,7 +1,10 @@
-import React, { useContext, useRef, useCallback } from 'react';
+import React, {
+  useContext, useRef, useCallback,
+} from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
 import ThemeContext from 'terra-theme-context';
+import { injectIntl } from 'react-intl';
 import * as KeyCode from 'keycode-js';
 import WorklistDataGridPropTypes from './proptypes/WorklistDataGridPropTypes';
 import './_elementPolyfill';
@@ -49,6 +52,11 @@ const propTypes = {
    * Number indicating the index of the column that represents row header. Index is 0 based and cannot exceed one less than the number of columns in the grid.
    */
   rowHeaderIndex: PropTypes.number,
+  /**
+   * @private
+   * Object containing intl APIs
+   */
+  intl: PropTypes.shape({ formatMessage: PropTypes.func }),
 };
 
 const defaultProps = {
@@ -62,6 +70,7 @@ function WorklistDataGrid(props) {
     ariaLabel,
     columns,
     rows,
+    intl,
   } = props;
 
   const focusedRow = useRef(0);
@@ -110,6 +119,7 @@ function WorklistDataGrid(props) {
       default:
         return;
     }
+
     if (nextRow >= grid.current.rows.length || nextCol >= grid.current.rows[0].cells.length) {
       event.preventDefault();
       return;
@@ -136,8 +146,9 @@ function WorklistDataGrid(props) {
     // Determine which cell was clicked. In the event that the user holds the mouse across multiple cells,
     // the originating cell is the clicked cell/active element.
     const clickedCell = event.target.closest('td,th') || document.activeElement.closest('td,th');
+
+    // Ensure that the cell is focusable and is a valid cell
     if (!clickedCell) {
-      // If anything other than a table data or table header cell is clicked, ignore the click.
       return;
     }
     // Remove Tab stop from previous cell in table that has focus and set it to the cell that was clicked.
@@ -152,15 +163,37 @@ function WorklistDataGrid(props) {
 
   const getCellData = (cell, cellColumnIndex) => {
     const tabIndex = { tabIndex: '-1' };
-    return props.rowHeaderIndex === cellColumnIndex ? (<th key={cellColumnIndex} {...tabIndex} className={cx('worklist-data-grid-row-header')}>{cell.content}</th>) : (<td key={cellColumnIndex} {...tabIndex} className={cx('worklist-data-grid-cell-data')}>{cell.content}</td>);
+
+    // Determine whether cell is a header or grid cell
+    const WorklistCellTag = props.rowHeaderIndex === cellColumnIndex ? 'th' : 'td';
+    let cellAriaLabel;
+
+    if (cell.isMasked) {
+      cellAriaLabel = intl.formatMessage({ id: 'Terra.worklistDataGrid.maskedCell' });
+    } else if (!cell.content) {
+      cellAriaLabel = intl.formatMessage({ id: 'Terra.worklistDataGrid.blank' });
+    }
+
+    return (
+      // Return worklist data grid cell component
+      <WorklistCellTag
+        key={cellColumnIndex}
+        {...tabIndex}
+        className={cx('worklist-data-grid-cell', { masked: cell.isMasked, selectable: !(cell.isMasked || cell.isSelectable === false), blank: !cell.content })}
+        aria-label={cellAriaLabel}
+      >
+        {!cell.isMasked && cell.content
+          && <div className={cx('cell-content')}>{cell.content}</div>}
+      </WorklistCellTag>
+    );
   };
 
-  const buildColumn = (columnData) => {
-    const width = columnData.width || props.columnWidth;
+  const buildColumn = (column) => {
+    const width = column.width || props.columnWidth;
     const height = props.columnHeaderHeight;
     return (
       /* eslint-disable react/forbid-dom-props */
-      <th key={columnData.id} className={cx('worklist-data-grid-column-header')} tabIndex="-1" style={{ width, height }}>{columnData.displayName}</th>
+      <th key={column.id} className={cx('worklist-data-grid-column-header', { selectable: !(column.isSelectable === false) })} tabIndex="-1" style={{ width, height }}>{column.displayName}</th>
     );
   };
 
@@ -168,7 +201,7 @@ function WorklistDataGrid(props) {
     if (allColumns?.length > 0) {
       return (
         <tr height={props.columnHeaderHeight}>
-          {allColumns.map(columnData => (buildColumn(columnData)))}
+          {allColumns.map(column => (buildColumn(column)))}
         </tr>
       );
     }
@@ -216,4 +249,4 @@ function WorklistDataGrid(props) {
 WorklistDataGrid.propTypes = propTypes;
 WorklistDataGrid.defaultProps = defaultProps;
 
-export default WorklistDataGrid;
+export default injectIntl(WorklistDataGrid);