+++ packages/terra-framework-docs/CHANGELOG.md
@@ -3,8 +3,11 @@
 ## Unreleased
 
 * Changed
-   * Updated example for `terra-popup`.
-   * Updated examples and tests for `terra-worklist-data-grid` to remove row height.
+  * Updated example for `terra-popup`.
+  * Updated examples and tests for `terra-worklist-data-grid` to remove row height.
+  * Updated examples and tests for `terra-worklist-data-grid` to remove row height
+  * Updated examples and tests for `terra-worklist-data-grid` for blank cell accessibility
+  * Updated examples and tests for `terra-worklist-data-grid` to show masked cells and hover states
 
 ## 1.23.0 - (June 12, 2023)
 
@@ -24,6 +27,7 @@
 
 * Added
   * Added documentation and tests for `terra-worklist-data-grid`.
+  * Added example and test for masked cell feature for `terra-worklist-data-grid`.
 
 ## 1.21.0 - (May 11, 2023)