+++ packages/terra-worklist-data-grid/CHANGELOG.md
@@ -3,6 +3,11 @@
 ## Unreleased
 * Changed
   * Removed ability to set variable row heights
+  * Updated styling on cell focus.
+* Added
+  * Added masked cell support.
+  * Added hover state for selectable cells.
+  * Added blank cell support.
 
 ## 0.1
-  * Initial stable release
+  * Initial stable release.